Fix gdb.rocm/show-info.exp regexps
Sometimes, the gdb.rocm/show-info.exp's "info queue" test misses a
queue.  E.g., GDB prints:

 (gdb) info queues
   Id   Target Id                Type         Read   Write  Size     Address
 * 1    AMDGPU Queue 1:1 (QID 1) HSA          1      3      1048576  0x00007ffee0a00000
   2    AMDGPU Queue 1:2 (QID 0) HSA          4      4      4096     0x00007ffff7faa000
   3    AMDGPU Queue 1:3 (QID 2) DMA                        1048576  0x00007ffeda200000

Yet, the test only sees two queues:

 (gdb) PASS: gdb.rocm/show-info.exp: info queue, queue 0
 (gdb) PASS: gdb.rocm/show-info.exp: info queue, queue 1
 (gdb) PASS: gdb.rocm/show-info.exp: info queues, no match

This doesn't currently cause a FAIL.  Instead simply we get one fewer
PASS than we should.  E.g., on a machine with gfx1030, we sometimes
get 16 passes, and other times, 17 passes.

The problem lies in the use of \\s* at the end of this:

 set table_header_regex "\\s+Target\\s+Id\\s+Type\\s+Read\\s+Write\\s+Size\\s+Address\\s*"

\\s is problematic because it not only consumes the space character
and tabs, but it also consumes carriage return (\r) and newline (\n),
among others.  I.e., if the expect buffer happens to contain:

  "... Address    \r\n"

then that \\s* will consume the \r\n as well.  If that happens, then
the regexp for the whole queue line:

     -re "\r\n(\\*?\\s*(\\d+)\\s..."

will not match for the first line, due to the leading missing (already
consumed) \r\n at the start.

A similar problem could happen in the "dbgapi properly configured"
test, as the '-re "Device' regexp ends with \\s.  If GDB ever outputs
something like "Device vvvv:dddd.rr\r\n" with no space before the
"\r\n", then that regexp will match the "\r", and then the '-re -wrap
""' regexp in the same test fails to match, because it expects
"\r\n$gdb_prompt ".

Basically, it's better just not to use "\\s", but instead match
"\[ \t\]" (space and tab).

That's the main change in this patch (there is no \\s left in the
testcase afterwards), but not the only change.

The next change is to make sure we consume the whole queue line.
Currently, the queue line regexp ends with:

   -re "...\\d+\\s+$::hex"

... and that $::hex can well do a partial match of a hex number.
E.g., with 0x012345678, that regexp matches 0x0 too, leaving
"12345678" in the buffer.  That is easy to reproduce with make
check-read1, except it doesn't cause a problem because later when we
match the saved queue line against what "info queue N" outputs, we
match it like so:

     gdb_test "info queue [lindex $queue_ids $i]" \
	"$table_header_regex\r\n[string_to_regexp [lindex $queue_info_lines $i]].*" \
        "info queue, queue $i"

Note that ".*" at the end of that regexp.  So if we earlier saved

   "....0x0"

This regexp becomes "....0x0.*" and it will still match against GDB
outputting "0x012345678".

I think that this is really a bug, though.  When we capture the "info
queues" lines originally, we should capture the whole line, and here
in the "info queue, queue $i" test, we should not have that ".*" at
the end.  That is addressed in the patch, too.

The next issue is that the test consumed too few "info queue" lines,
but nothing FAILed, nothing noticed the issue.  That suggests a
missing test to me.  Fix that by adding a loop that makes sure that
GDB outputs queues in order starting at queue 1.  If we had this test
before it would have exposed the issue with sometimes not being able
to extract the first queue.

Lastly, since queues are output in order, and we verify their number,
we can simplify the way we come up with the first invalid queue id, as
it is always the next after the last queue we saw.

Change-Id: I37cc9d47f95b3b8ce58d988f0e487e6156c494be
(cherry picked from commit 19c55b3be0b7d341e0ef8b0e96aa55be1f1aa23b)