Don't rely on rocm_agent_enumerator
rocm_agent_enumerator does not exist on all HIP supported platforms,
so instead of calling rocm_agent_enumerator, compile our own
replacement HIP program that lists GPU devices (only).

With this, we no longer need to check if we have a working hipcc
compiler available, as compiling the rocm_agent_enumerator replacement
HIP program achieves the same goal.  In turn this means that the cost
of having this replacement is virtually zero.

Change-Id: I45262f2fe24455ade6075eded2cd7df03979c6cc