testsuite/lib/set_unbuffered_mode.c: Use constructor priority
On native Windows, we link in testsuite/lib/set_unbuffered_mode.c to
all executables.  That file has a constructor function that disables
output buffering.  We want that constructor to run before the
constructors of global C++ objects, in case those objects print to
output.

Currently, that is done by relying on the link order of object files
on the link line affecting the global ctor order.  On Windows, with
gcc and clang targeting the GNU ABI, constructors are run in reverse
link order.  I.e., the constructors from the last objfile on the link
line run first.

That is not true if we compile with clang targeting the MSVC ABI,
though.  There, global ctors run on link line order.  Like on Linux.

This could instead be handled by setting a constructor priority.
Constructors with a priority run before the global C++ objects'
constructors.  The GCC feature to allow specifying a constructor with
a priority only appeared in GCC 4.3, released in 2008, and I added
this set_unbuffered_mode feature to GDB also in 2008, so back then it
was too early to rely on the priority feature.

It's 2025 now, and I think we can safely assume nobody is testing on
Windows with such an old compiler.

So this patch gives an explicit priority to the
testsuite/lib/set_unbuffered_mode.c constructor, and tweaks the
comments in testsuite/lib/gdb.exp to no longer talk about controlling
ctor order with link order.

Change-Id: Ia5f8972f2ef45f070e63cb0a1b0aa9918957ea8e
(cherry picked from commit c475a81171a04a0452f9bc86d20b54668749ac77)