gdb/utils.c: refactor strncmp_iw_with_mode_tests unit test
This change is a preparatory step towards adding a dialect of C++
language to GDB.

There are around ~30 checks in this test that are meant for
language_cplus.  This change parameterises that section by
introducing a lambda, so one can easily use it for other
languages as well.  These "other languages" are all supposed
to be extensions of C++.

While at it, I used some macros to make the lengths of the
lines shorter and in return make the test clearer.

Change-Id: Ibcf008fa34559bcd01ab00cb1d4b316e3f09681f