gdb/language: add "is_cplus_dialect ()" check
This change is a preparatory step towards adding a dialect of C++
language to GDB.

There are many instances in the code that check if the "language"
in the context is a C++ language (or not):

  ... language == language_cplus ...
  ... language != language_cplus ...

These checks can easily explode if dialects of C++ language are
to be supported.  Instead, this patch adds an inline function to
gdb/language.h, that, for now, checks against "language_cplus".
Then the patch continues to replace all those checking instances
with the inline function calls:

  ...  is_cplus_dialect (language) ...
  ... !is_cplus_dialect (language) ...

Later, if the support for a new dialect of C++ language is to be
added, mostly the "is_cplus_dialect ()" is going to be updated
(along with the switch cases sprinkled in the code) to cover all
the checks.

Change-Id: I23bb9f77eece4d2c85891bf7e4443b7a8a193c7a