gdb/testsuite/unittests: refactor lookup_name_info test
This change is a preparatory step towards adding a dialect of C++
language to GDB.

There is a section in the test that performs some checks for the
C++ language (language_cplus).  This change parameterises that
section by introducing a lambda, so one can easily use it for
other languages as well.  These "other languages" are all supposed
to be extensions of C++.

Change-Id: I643c688ad3bbd26234253b34b7a5fff4c41e6bf6
(cherry picked from commit 210b644b78d294773880163950fb2c3f313c9e35)