rocm_code_object_stream_memory::read crash, add bounds checking
It is possible for a memory:// code object to be corrupted by some
rogue pointers or some such, since it is all in memory.  One likely
reason for corruption is that the program reuses the memory where the
code object's image is stored.  GDB creates a snapshot of code objects
when it first sees them, to try to work around such programs, but that
doesn't help the attach case.

Usually, when you attach to an already-running process that has code
object images with corrupted ELF headers, you see something like:

 Error while mapping shared library sections:
 `memory://118381#offset=0x7d7613311f70&size=29000': not in executable format: file format not recognized

We observed a case of such corruption crashing GDB, though.  When it
happened, the ELF's header was broken such that the e_shoff field
pointed outside the bounds of the ELF file (among other broken
fields).

E.g.:

 (top-gdb) p pathname
 $1 = 0x55ec11cff0a0 "memory://44039#offset=0x7da2e363b720&size=16600"

And in elfcode.h:elf_object_p:

 (top-gdb) p /x *i_ehdrp
 $2 = {e_ident = {0x7f, 0x45, 0x4c, 0x46, 0x2, 0x1, 0x1, 0x40, 0xa5, 0x0, 0x0, 0x0, 0x0, 0x0, 0x0, 0x0}, e_entry = 0x7da26c862090, e_phoff = 0x5, e_shoff = 0x7da2e3351fe0,
  e_version = 0x7f82, e_flags = 0xe3335f80, e_type = 0xe620, e_machine = 0x6b8, e_ehsize = 0x7da2, e_phentsize = 0x0, e_phnum = 0xd, e_shentsize = 0x0, e_shnum = 0x0, e_shs
 trndx = 0x0}
 (top-gdb)

Here's the crashing backtrace:

  Thread 1 "rocgdb" received signal SIGSEGV, Segmentation fault.
  __memmove_avx512_unaligned_erms () at ../sysdeps/x86_64/multiarch/memmove-vec-unaligned-erms.S:317
  317     ../sysdeps/x86_64/multiarch/memmove-vec-unaligned-erms.S: No such file or directory.
  (top-gdb) bt
  #0  __memmove_avx512_unaligned_erms () at ../sysdeps/x86_64/multiarch/memmove-vec-unaligned-erms.S:317
  #1  0x000055ebbb1151cd in memcpy (__len=64, __src=<optimized out>, __dest=<optimized out>) at /usr/include/x86_64-linux-gnu/bits/string_fortified.h:29
  #2  (anonymous namespace)::rocm_code_object_stream_memory::read (this=<optimized out>, buf=<optimized out>, size=64, offset=<optimized out>) at src/ROCgdb/gdb/solib-rocm.c:443
  #3  0x000055ebbaf61f34 in operator() (__closure=<optimized out>) at src/ROCgdb/gdb/gdb_bfd.c:963
  #4  catch_exceptions<long int, -1, gdb_bfd_openr_iovec(char const*, char const*, gdb_iovec_opener_ftype)::<lambda(bfd*, void*, void*, file_ptr, file_ptr)>::<lambda()> > (f=...) at src/ROCgdb/gdb/event-top.h:96
  #5  operator() (__closure=0x0, offset=<optimized out>, nbytes=<optimized out>, buf=<optimized out>, stream=<optimized out>, nbfd=<optimized out>) at src/ROCgdb/gdb/gdb_bfd.c:961
  #6  _FUN () at src/ROCgdb/gdb/gdb_bfd.c:971
  #7  0x000055ebbb2793f9 in opncls_bread (abfd=<optimized out>, buf=<optimized out>, nbytes=<optimized out>) at src/ROCgdb/bfd/opncls.c:644
  #8  0x000055ebbb2727ef in bfd_read (ptr=ptr@entry=0x7ffe98d43d50, size=size@entry=64, abfd=<optimized out>, abfd@entry=0x55ec11d04fb0) at src/ROCgdb/bfd/bfdio.c:369
  #9  0x000055ebbb2934e7 in bfd_elf64_object_p (abfd=0x55ec11d04fb0) at src/ROCgdb/bfd/elfcode.h:645
  #10 0x000055ebbb276d34 in bfd_check_format_matches (abfd=0x55ec11d04fb0, format=<optimized out>, matching=<optimized out>) at src/ROCgdb/bfd/format.c:510
  #11 0x000055ebbb115252 in rocm_solib_bfd_open (pathname=0x55ec11cff0a0 "memory://44039#offset=0x7da2e363b720&size=16600") at src/ROCgdb/gdb/solib-rocm.c:617
  #12 0x000055ebbb122a36 in solib_map_sections (so=...) at src/ROCgdb/gdb/solib.c:548
  ...

Dumping that particularly broken memory ELF to a file and then loading
that file back into GDB (with add-symbol-file to trigger the same code
paths), GDB rejects it without crashing.  That is because there's a
bfd_seek call to e_shoff (to 0x7da2e3351fe0 in that example) that
turns into an lseek call on a real file, which fails.  Here, in
bfd/elfcode.h:elf_object_p:

      /* Seek to the section header table in the file.  */
      if (bfd_seek (abfd, where, SEEK_SET) != 0)
	goto got_no_match;               <<<<<<<<<<<<<< we hit this

When the ELF is in memory, that bfd_seek call always succeeds, as it
just manipulates the BFD's 'where' offset.  We then reach the bfd_read
call just after:

      /* Read the first section header at index 0, and convert to internal
	 form.  */
      if (bfd_read (&x_shdr, sizeof x_shdr, abfd) != sizeof (x_shdr))
	goto got_no_match;

which takes us to rocm_code_object_stream_memory::read [*].  This
function assumes the 'offset' argument is within bounds of the
m_objfile_image buffer, which it isn't, and thus we crash in the
memcpy call.

[*] - This method is slighly misnamed IMHO.  The BFD side calls the
callback 'pread', which is more accurate, as this has an explicit
offset argument.

Fix it by making rocm_code_object_stream_memory::read bounds check its
'offset' argument.

The 'size' parameter is also signed, but if we look up the call stack
before we get to rocm_code_object_stream_memory::read, it is unsigned
there:

  bfd_size_type bfd_read (void *ptr, bfd_size_type size, bfd *abfd);

IMHO, the abfd->iovec->bread interface should take an unsigned size
too, modelling 'pread', but I'm not changing bfd in this patch.  So
for now, also error out if we see a negative size request, as we'd
behave incorrectly if we saw one.

Tweak the existing gdb.rocm/snapshot-objfile-on-load.exp testcase to
exercise this scenario too, by:

 - Adding a new command line option to tell the test program how to
   scramble module image memory.

 - Adding a new "shoff" mode, that scrambles section offsets instead
   of zeroing the whole module image buffer.

 - Forcing GDB to reload shared libraries after the module image
   buffer has been corrupted, thus discarding the pristine local copy
   GDB used to have, and forcing GDB to parse the broken ELF.

Change-Id: Ie72ef895802057591d6e06cd48f85edf1c177993