gdb/language: add container for built-in symbols
Add a vector to "language_arch_info" to keep track of built-in symbols.
These built-in symbols are context variables defined by the programming
language and known to the compiler.  In this patch, there's no looking
up for these symbols yet.  In the next patches, the HIP language will
be using them to look for its built-in symbols.

Change-Id: Ib3b22ef4c05a5037ef48b253715cdd257ba5f365