gdb/language/hip: add built-in symbol lookup
Introduce a "lookup_symbol_nonlocal ()" method to "hip_language" that
looks up built-in symbols before moving onto global symbols.

built-in symbols are not conventional variables that can be easily
encoded in debug information.  They are some sort of context variables
that the HIP compiler knows about and emits instructions into the
device code to retrieve them.

This patch teaches the "hip_language" about those symbols and adds
target hooks in place for retrieving them.  Here's the list of those
built-in symbols:

  threadIdx  -- position of the current lane in the HIP block
  blockIdx   -- position of the current block in the HIP grid
  blockDim   -- sizes of the HIP block in x,y,z dimensions
  gridDim    -- sizes of the HIP grid in number of blocks (x,y,z)
  warpSize   -- number of lanes in current wave

The added target hooks are:

lane_workgroup_pos:
Given a thread info and lane number this hook shall return the
coordinates of the work item this lane is working on.

workgroup_grid_pos:
Given a thread info this hook returns the grid coordinates of
the work-group that contains that thread.

workgroup_sizes:
Returns the sizes in the x,y,z dimensions of the work-group that
contains a given thread.

grid_sizes:
Returns the sizes in the x,y,z dimensions of the grid that contains
a given thread.

wave_size:
Given a thread info this hook returns the wave size of the platform
that the thread belongs to.

For these new hooks, new data types are added to target.h:

  using vec3_t = std::array<T, 3>;
  using vec3_u32_t = vec3_t<uint32_t>;
  using opt_vec3_u32_t = std::optional<vec3_u32_t>;
  using opt_size_t = std::optional<size_t>;

Co-Authored-By: Simon Marchi <simon.marchi@efficios.com>
Co-Authored-By: Pedro Alves <pedro@palves.net>
Change-Id: I134b92fffaf2e03a8e72d516d232cc7a16a59da0