gdb/solib-rocm: assert that host ops isn't rocm_solib_ops
In some fork cases, the rocm_solib_target_inferior_created observer gets
called while the child inferior (passed as a parameter) already has a
rocm_solib_ops installed.  Since we unconditionnally wrap the existing
solib_ops with a new rocm_solib_ops, we can end up with a chain of
multiple rocm_solib_ops, like:

  rocm_solib_ops -> rocm_solib_ops -> svr4_solib_ops

I don't think it is that harmful (unless the process does a ton of forks
and the rocm_solib_ops accumulate), but it for sure useless.  Add an
assert for this in the rocm_solib_ops constructor, which reveals the
cases where this happens.

Change rocm_solib_target_inferior_created to only create and set a
rocm_solib_ops if the inferior's pspace doesn't already have one.

I would have liked to be able to detect exactly whether we need to set a
new rocm_solib_ops, based on the combination of fork/vfork,
follow-fork-mode and detach-on-fork.  I am a bit afraid of some cases
where we would erronesouly leave a rocm_solib_ops in place, and with
this code we wouldn't catch it.  But we don't have all the necessary
info at this point in the inferior_created observer.  So, just checking
if the pspace already has a rocm_solib_ops it is.

Change-Id: I2e76d111e96f1e01b6799b04da9cdd6f6e2984c9