Introduce allow_rocm_core_tests and use it
Core generation is not supported on Windows, so introduce a new
allow_rocm_core_tests procedure and use it in core-related testcases.

gdb.rocm/core-no-read-special-files.exp, in addition to being about
core files, it's about /dev/kfd, which is Linux-specific.  So I added
a require linux to the file too, in addition to allow_rocm_core_tests,
in case core files are ever supported on non-Linux hosts.

Change-Id: I70a74b015598d082d9fa04da2a4386eee6681080