gdb/testsuite/gdb.rocm: Add missing includes
Most of the testcases in the gdb.rocm directory have a CHECK macro that
uses the fprintf(3) and exit(3), but do not include explicitly stdio.h
nor stdlib.h.  Instead, those testcases implicitly rely on
hip/hip_runtime.h to transitively include those files.

This patch makes it so we explicly include headers we rely on.

Change-Id: I61d21c5af0646ed91ef96870928801b43da598ca