gdb/testsuite: Fix tests violating absolute path on clean_restart restriction
Upstream now issues errors for tests violating the rule that clean_restart
should not be called with an absolute path.  These tend not to catch the eye
as they are summarized as UNRESOLVED.

Fix a couple of these in gdb.base and an obscure one that is seldom executed
in gdb.perf.