gdb.rocm/lane-execution.exp: Don't expect forward slash
On Windows, we currently get:

 bt
 #0  foo (val=1) at C:/rocgdb/src.cascais-rsync/gdb/testsuite/gdb.rocm\lane-execution.cpp:67
 #1  0x000000030000ac08 in lane_pc_test (gid=1, in=0x300005000, out=0x300004000) at C:/rocgdb/src/gdb/testsuite/gdb.rocm\lane-execution.cpp:100
 #2  0x000000030000ceb4 in kernel (in=0x300005000, out=0x300004000) at C:/rocgdb/src/gdb/testsuite/gdb.rocm\lane-execution.cpp:120
 (gdb) FAIL: gdb.rocm/lane-execution.exp: step: then: bt on then lane

The issue is the "\" in "\lane-execution.cpp".  The .exp is expecting
"/lane-execution.cpp".

Remove the "/" expectation, which would also be incorrect when
remote-host testing, in which case the filename can sometimes just be
${srcfile} with no leading path.

Change-Id: I8087d7f49b9aaba6c1e9162271a32af3af7fcade
(cherry picked from commit 0f54f2bc1dcc52d462b2c87fb7877542bd93853c)