gdb/testsuite: New runtime copy code object on load
New runtime makes a copy of a memory:// code object when loading it,
reporting the address of that copy to the debugger.  With this change,
an application cannot override the code object that the debugger might
try to read later.

This is a good thing, but it breaks a test where we would check that GDB
would not crash if the original memory is modified by the application.
Since we now see a pristine version of the code object, we do not see an
error message anymore.

Adjust the gdb/testsuite/gdb.rocm/snapshot-objfile-on-load.exp to not
look for the error message, and just ensure that GDB handled everything
normally.

Change-Id: Ibba05b4daceb89b898be975f719a25e1db773eb0