gdb/language: add "strip_cplus_dialect ()"
This is yet another preparatory change before adding a dialect of
C++ language to GDB.

It is meant for pieces of code that have a pattern similar to:

  switch (lang)
    {
    ...
    case language_cplus:
      do_something ();
      break;
    ...
    }

When adding a new language that is a dialect of C++, that would
require to add new "case"s for it:

  switch (lang)
    {
    ...
    case language_cplus:
    case language_a_cplus_dialect:
      do_something ();
      break;
    ...
    }

Instead, we have opted for the introduction of a new function called
"strip_cplus_dialect ()" that returns "cplus_language" for all the
dialects of C++, including C++ itself.  With that in place, code
snippets that look like the above pattern are changed to:

  switch (strip_cplus_dialect (lang))
    {
    ...
    case language_cplus:
      do_something ();
      break;
    ...
    }

Addition of new C++ dialects this way, no more requires adding new
"cases", as long as the "is_cplus_dialect ()" (the function used
by "strip_cplus_dialect ()") is updated.

Change-Id: If50d383e22d39049850c7b1238fd4714765bdd96