gdb.rocm/lane-execution.exp: Handle step out of else
On some systems, stepping out of the 'else' branch brings you to the
'then' branch directly.  On other systems, that stops at the 'if'
condition first, and so we need another step to get to the 'then'
branch.

Make the gdb.rocm/lane-execution.exp test handle the latter kind of
systems.

Change-Id: I3ae599952e567ccacd4bdfd59debdc5b44d473e2
commit-id:bd246cd4