Fix device index in gdb.rocm/addr-bp-gpu-no-deb-info.exp
On a system with a single GPU, I spotted this test failing. The environment
variable ROCR_VISIBLE_DEVICES can be a bit misleading, as it actually expects
a list of device id's as opposed to a literal number of devices to make
visible.

This test sets it to 1, which is the second GPU on the system. As a result,
systems with a single GPU will have no visible GPU's and the test will FAIL.

Set ROCR_VISIBLE_DEVICES to 0 to make use of the first GPU on the system.

Approved-By: Lancelot Six <lancelot.six@amd.com> (AMDGPU)
(cherry picked from commit c1c173422d19d1c38aa80fbf689e054a4a19c708)