gdb/charset: Use UCS-4 if it is a better alternative than UTF-32
Some systems (azurelinux for example) do not seem to have support for
UTF-32 out of the box.  We can use UCS-4 istead.

This patch proposes to try to detect at runtime which of the two
encoding is available, and select it.  If neither is available, we keep
using UTF-32 as today, and expect code using it to fail gracefully.

Change-Id: Ib62ae5143b14680fbf1acf3f1451ec3eac4e27a2
Bug: SWDEV-562503