Fix FAILs in gdb.python/py-pp-maint.exp
I spotted some FAILs in this test related to the assumption that gdb
starts up free of Python pretty-printers. However this is not the case,
and we could have builtin pretty-printers shipped with gdb that get
auto-loaded.

This patch detects how many preloaded pretty-printers we have and accounts
for those when running the actual test.

Bug: SWDEV-527307