gdb: refuse to produce .gdb_index when skeletonless type units are present
Running test gdb.dwarf2/fission-with-type-unit.exp with the
cc-with-gdb-index target board fails with:

    (gdb) maint expand-symtabs
    /home/simark/src/binutils-gdb/gdb/dwarf2/read.c:3064: internal-error: cutu_reader: Assertion `sig_type->signature == cu->header.signature' failed.
    A problem internal to GDB has been detected,
    further debugging may prove unreliable.
    ----- Backtrace -----
    FAIL: gdb.dwarf2/fission-with-type-unit.exp: maint expand-symtabs (GDB internal error)

This is a consequence of .gdb_index not supporting skeletonless type
units in .dwo files.  That is, type units in .dwo files that don't have
a corresponding skeleton (or stub) in the main file.

For context: in DWARF 4, gcc 4.x used to create skeletons for type units
in .dwo files, but subsequent versions don't.  DWARF 5 doesn't have
support for type unit skeletons at all.  So skeletons for type units are
mostly a historical curiosity at this point, the norm is to not have
them.

Here's what leads up to the crash.  First, this is what is in the main
file's .debug_info section (the first and last CUs are dummy CUs added
by the testsuite):

      Compilation Unit @ offset 0:
       Length:        0x8 (32-bit)
       Version:       4
       Abbrev Offset: 0
       Pointer Size:  8
     <0><b>: Abbrev Number: 1 (DW_TAG_compile_unit)

      Compilation Unit @ offset 0xc:
       Length:        0x15 (32-bit)
       Version:       5
       Unit Type:     DW_UT_skeleton (4)
       Abbrev Offset: 0x6
       Pointer Size:  8
       DWO ID:        0xf00d
     <0><20>: Abbrev Number: 1 (DW_TAG_compile_unit)
        <21>   DW_AT_dwo_name    : (strp) (offset: 0): fission-with-type-unit-dw.dwo

      Compilation Unit @ offset 0x25:
       Length:        0x8 (32-bit)
       Version:       4
       Abbrev Offset: 0xe
       Pointer Size:  8
     <0><30>: Abbrev Number: 1 (DW_TAG_compile_unit)

And here is what is in the fission-with-type-unit-dw.dwo file (one TU
and the CU):

    Contents of the .debug_info.dwo section:

      Compilation Unit @ offset 0:
       Length:        0x1d (32-bit)
       Version:       5
       Unit Type:     DW_UT_type (2)
       Abbrev Offset: 0
       Pointer Size:  8
       Signature:     0xcafe
       Type Offset:   0x19
     <0><18>: Abbrev Number: 1 (DW_TAG_type_unit)
     <1><19>: Abbrev Number: 2 (DW_TAG_base_type)
        <1a>   DW_AT_byte_size   : (sdata) 4
        <1b>   DW_AT_encoding    : (sdata) 5        (signed)
        <1c>   DW_AT_name        : (string) int
     <1><20>: Abbrev Number: 0

    Contents of the .debug_info.dwo section:

      Compilation Unit @ offset 0:
       Length:        0x2d (32-bit)
       Version:       5
       Unit Type:     DW_UT_split_compile (5)
       Abbrev Offset: 0
       Pointer Size:  8
       DWO ID:        0xf00d
     <0><14>: Abbrev Number: 3 (DW_TAG_compile_unit)
     <1><15>: Abbrev Number: 4 (DW_TAG_base_type)
        <16>   DW_AT_byte_size   : (sdata) 4
        <17>   DW_AT_encoding    : (sdata) 5        (signed)
        <18>   DW_AT_name        : (string) int
     <1><1c>: Abbrev Number: 5 (DW_TAG_variable)
        <1d>   DW_AT_name        : (string) global_var
        <28>   DW_AT_type        : (ref4) <0x15>, int
        <2c>   DW_AT_location    : (exprloc) 3 byte block: 8 c 9f   (DW_OP_const1u: 12; DW_OP_stack_value)
     <1><30>: Abbrev Number: 0

After loading the above in GDB, here is what is in GDB's mind (contents
of dwarf2_per_bfd::all_units):

 - CU at offset 0x0  of .debug_info     in fission-with-type-unit -- dummy
 - CU at offset 0xc  of .debug_info     in fission-with-type-unit
 - CU at offset 0x25 of .debug_info     in fission-with-type-unit -- dummy
 - TU at offset 0x0  of .debug_info.dwo in fission-with-type-unit-dw.dwo

This is correct.  Then, this is the generated .gdb_index:

    Contents of the .gdb_index section:

    Version 9

    CU table:
    [  0] 0 - 0xb     -- dummy
    [  1] 0xc - 0x24
    [  2] 0x25 - 0x30 -- dummy

    TU table:
    [  0] 0 0x19 000000000000cafe

    Address table:

    Symbol table:
    [  3] global_var: 1 [static, variable]
    [754] int: 1 [static, type]

    Shortcut table:
    Language of main: unknown: 0
    Name of main: <unknown>

The TU table says that there exists a TU at offset 0.  Unfortunately,
there is no way for a reader of that index to know that this TU is
really in a .dwo file, not in the main file.  So when GDB loads this
index back (creating dwarf2_per_bfd::all_units from .gdb_index this
time, rather than walking the debug info), this is what is in its mind:

 - CU at offset 0x0  of .debug_info in fission-with-type-unit -- dummy
 - TU at offset 0x0  of .debug_info in fission-with-type-unit
 - CU at offset 0xc  of .debug_info in fission-with-type-unit
 - CU at offset 0x25 of .debug_info in fission-with-type-unit -- dummy

GDB now incorrectly believes there's a TU at offset 0 of .debug_info in
the main file, which is wrong.  When trying to expand that TU with
"maint expand-symtabs", we're not really reading the TU, so we hit the
assert checking that the signature in the TU header matches what was
given by the index.

The .debug_names format has a way to list the TUs found in the .dwo
files, called the "foreign TU list" (see section 6.1.1.2 "Structure of
the Name Index" of DWARF 5).  That list only includes the signature of
the type, and there is some capability to figure out which .dwo file
contains that type unit.  The .gdb_index format does not have something
like that.  We could try to retrofit such a feature in the .gdb_index
format, but I think we prefer to put our efforts on the standard
.debug_names format.

To avoid producing a misleading index like shown above, I propose to
make GDB refuse to produce an index if there exists a skeletonless type
unit.  This patch implements that by looking at section of all
signatured_types.  If the containing section ends with .dwo, then this
is a skeletonless type unit.

As a reminder: if a unit has a skeleton, the dwarf2_per_cu section will
point at the skeleton section, in the main file.  If the unit does not
have a skeleton, the dwarf2_per_cu section will point at the section in
the .dwo file.  All .dwo section names end with ".dwo".

Add a "endswith" utils function to help with that.

With this patch, running the gdb.dwarf/fission-with-type-unit.exp leads
to a compilation failure:

    gdb compile failed, Error while writing index for `/home/smarchi/build/binutils-gdb/gdb/testsuite/outputs/gdb.dwarf2/fission-with-type-unit/.tmp/fission-with-type-unit': Found skeletonless type units, unable to produce .gdb_index.  Consider using .debug_names instead.

... which makes the test "untested".

Add a new test, gdb.dwarf2/gdb-index-skeletonless-tu.exp, to verify the new error path.

Change-Id: I1e2e0204c9c2b48763aa99ce63521ae4a5262b22
Reviewed-By: Eli Zaretskii <eliz@gnu.org>
Approved-By: Tom Tromey <tom@tromey.com>